+++ Documentation/Changelog.md
@@ -2,6 +2,10 @@
 
 ## 2026
 
+### Gapotchenko.FX 2026.5 (Beta)
+
+Release date: not released yet
+
 ### Gapotchenko.FX 2026.4 (Beta)
 
 Release date: February 4, 2026